Move initialization/shutdown test to integration test
With new support for sending `initialized()` too
